Problem with 816-tcc
The tinycc, or 816-tcc, is a very widely copy-and-save repository. The workability of this code base into "project melchior" is not going well. Actually, it is the only library that does not compile! That's quite something that they all worked out almost immediately. So, I was anticipating this sort of thing with every one of these old C libs. There will be another... project "something" ... very soon. This treachery will not stand!